Move coerce_to_dataframe
