+++ ibis/backends/pandas/execution/util.py
@@ -9,6 +9,7 @@ import ibis.common.exceptions as com
 import ibis.util
 from ibis.expr import operations as ops
 from ibis.expr import types as ir
+from ibis.expr.schema import coerce_to_dataframe
 from ibis.expr.scope import Scope
 
 from ..core import execute
@@ -107,12 +108,16 @@ def coerce_to_output(
     result_name = getattr(expr, '_name', None)
 
     if isinstance(expr, (ir.DestructColumn, ir.StructColumn)):
-        return ibis.util.coerce_to_dataframe(result, expr.type().names)
+        return coerce_to_dataframe(
+            result, expr.type().names, expr.type().types
+        )
     elif isinstance(expr, (ir.DestructScalar, ir.StructScalar)):
         # Here there are two cases, if this is groupby aggregate,
         # then the result e a Series of tuple/list, or
         # if this is non grouped aggregate, then the result
-        return ibis.util.coerce_to_dataframe(result, expr.type().names)
+        return coerce_to_dataframe(
+            result, expr.type().names, expr.type().types
+        )
     elif isinstance(result, pd.Series):
         return result.rename(result_name)
     elif isinstance(expr.op(), ops.Reduction):
